Changed and fixed some things.
Changed GetChildren to DoForEveryChild, so instead of 2 loops, it will do things in 1 loop.

Changed tkinter to PyQt to use OpenGL and hardware acceleration.

Deleted version showing in console, because version is no more in offsets api.

Changed gravity changer to camera zoomer when right button mouse is clicked, because roblox patched gravity changing method and idk how to fix it.